+++ pkg/commands/integration/doc.go
@@ -1,0 +1,5 @@
+// Package integration contains commands to inspect and manipulate Fastly
+// notification integrations.
+//
+// https://www.fastly.com/documentation/reference/api/observability/notifications/
+package integration