+++ pkg/commands/integration/jiraissue/doc.go
@@ -1,0 +1,3 @@
+// Package jiraissue contains commands to manage Jira Issue notification
+// integrations.
+package jiraissue