+++ pkg/commands/integration/splunkoncall/doc.go
@@ -1,0 +1,3 @@
+// Package splunkoncall contains commands to manage Splunk On-Call
+// notification integrations.
+package splunkoncall